backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
sqlalchemy>=2.0.30
psycopg2-binary==2.9.9
reportlab>=4.0
pandas>=2.0.0
pdfplumber==0.11.6
python-multipart>=0.0.9
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
sqlalchemy>=2.0.30
psycopg2-binary==2.9.9
reportlab>=4.0
pandas>=2.0.0
pdfplumber==0.11.6
python-multipart>=0.0.9
python-dotenv
passlib[bcrypt]
bcrypt==3.2.2
python-jose[cryptography]
tzdata
requests
slowapi>=0.1.9
python-json-logger>=2.0.0